Support more Python types
